pyproject.toml
[tool.poetry]
name = "metac-bot-template"
version = "0.1.0"
description = ""
authors = ["Vasile Popescu <elisescu@elisescu.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
python-decouple = "^3.8"
requests = "^2.32.3"
asknews = "^0.11.2"
numpy = "^2.3.0"
openai = "^1.57.4"
python-dotenv = "^1.0.1"
forecasting-tools = "0.2.54"
litellm = {extras = ["proxy"], version = "^1.0.0"}
aiohttp = "^3.12.15"
yfinance = "^0.2.48"
fredapi = "^0.5.2"
google-genai = "^1.74.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"

pre-commit = "^4.3.0"
ruff = "^0.15.1"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.uv]
# Supply chain safety: ignore packages published in the last 7 days.
exclude-newer = "1 week"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
